==> scp/2-model_training.R <==


# RUN ---------------------------------------------------------------------

sys.source("scp/2-model_training-europe.R", envir = new.env())
sys.source("scp/2-model_training-global.R", envir = new.env())
sys.source("scp/2-model_training-italy.R", envir = new.env())
