==> scp/3-model_evaluation.R <==


# RUN ---------------------------------------------------------------------

sys.source("scp/3-model_evaluation-europe.R", envir = env())
sys.source("scp/3-model_evaluation-global.R", envir = env())
sys.source("scp/3-model_evaluation-italy.R", envir = env())
